+++ src/main/java/com/skrra/atmosphereplus/config/AtmosphereConfig.java
@@ -27,8 +27,10 @@ public class AtmosphereConfig {
     public boolean fogOverride = false;
     public float fogDistance = 1.0f;
     public float fogDensity = 1.0f;
+    public boolean submersionFogOff = false;
 
     public float particleAmount = 1.0f;
+    public boolean lowFire = false;
 
     public boolean cloudOverride = false;
     public String cloudMode = "SERVER";